+++ js/wasm/src/webworker/requirements.ts
@@ -1,3 +1,5 @@
+import type { PyodideInterface } from "pyodide";
+
 export function verifyRequirements(requirements: string[]): void {
 	requirements.forEach((req) => {
 		let url: URL;
@@ -15,4 +17,42 @@ export function verifyRequirements(requirements: string[]): void {
 			);
 		}
 	});
+}
+
+function isPlotly6(pyodide: PyodideInterface, requirement: string): boolean {
+	const pyRequirement = pyodide.pyimport("packaging.requirements.Requirement");
+	try {
+		const reqObj = pyRequirement(requirement);
+		return reqObj.name === "plotly" && reqObj.specifier.contains("6");
+	} catch (error) {
+		return false;
+	}
+}
+
+function isAltair(pyodide: PyodideInterface, requirement: string): boolean {
+	const pyRequirement = pyodide.pyimport("packaging.requirements.Requirement");
+	try {
+		const reqObj = pyRequirement(requirement);
+		return reqObj.name === "altair";
+	} catch (error) {
+		return false;
+	}
+}
+
+export function patchRequirements(pyodide: PyodideInterface, requirements: string[]): string[] {
+	// XXX: `micropip` sometimes doesn't resolve the dependency version correctly.
+	// So we explicitly specify the version here for some packages.
+
+	if (requirements.some((req) => isAltair(pyodide, req))) {
+		// Plotly 6.x doesn't work work Altair on Pyodide 0.27.2.
+		// Ref: https://github.com/gradio-app/gradio/issues/10458
+		return requirements.map((req) => {
+			if (isPlotly6(pyodide, req)) {
+				return `plotly==5.*`;
+			}
+			return req;
+		});
+	}
+
+	return requirements;
 }
